Refactor Maker state machine
