+++ maker/package.json
@@ -18,6 +18,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@paybase/machine": "^1.0.12",
     "comit-sdk": "^0.13.0",
     "concurrently": "^5.1.0",
     "fastify": "^2.0.0",
